+++ src/lib/mission/defaults.ts
