+++ src/lib/mission/fns.ts
