+++ src/util/context/cloudWaypointProvider.tsx
@@ -2,7 +2,7 @@
 import { useState } from 'react';
 import { waypointContext } from './WaypointContext';
 import { Tool } from '@/types/tools'
-import { Mission } from '@/lib/waypoints/waypointCollection';
+import { Mission } from "@/lib/mission/mission";
 import { Node } from '@/types/waypoints';
 
 type Props = {
